Consolidate / Port oauth addon tests
