+++ addons/base/models.py
@@ -216,13 +216,13 @@ class BaseOAuthUserSettings(BaseUserSettings):
         """
         for node in self.get_nodes_with_oauth_grants(external_account):
             try:
-                addon_settings = node.get_addon(external_account.provider, deleted=True).deauthorize(auth=auth)
+                node.get_addon(external_account.provider, deleted=True).deauthorize(auth=auth)
             except AttributeError:
                 # No associated addon settings despite oauth grant
                 pass
 
         if external_account.osfuser_set.count() == 1 and \
-                external_account.osfuser_set.filter(osfuser=auth.user).count() == 1:
+                external_account.osfuser_set.filter(id=auth.user.id).exists():
             # Only this user is using the account, so revoke remote access as well.
             self.revoke_remote_oauth_access(external_account)
